projects/requirements.txt
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
Django==2.2
django-allauth==0.41.0
django-default-imagefield==0.5.5
django-filter==2.2.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
httpie==2.0.0
idna==2.8
oauthlib==3.1.0
Pillow==7.0.0
Pygments==2.5.2
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
